requirements.txt
# Core packages
datasets==2.16.1
huggingface-hub==0.20.3
transformers==4.37.2

# PyTorch for Apple Silicon M4
torch==2.1.2

# Vector store and embeddings
sentence-transformers==2.2.2
transformers==4.37.2
huggingface-hub==0.20.3
tokenizers==0.15.2
faiss-cpu==1.7.4
chromadb==0.4.22
hf-xet==1.2.0

# LLM integration
langchain==0.1.6
langchain-community==0.0.19
groq==1.0.0

# Evaluation and analysis
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.5
seaborn==0.13.1

# Deployment
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1

# Sparse Retrieval
rank-bm25==0.2.2
openai==2.1.4

gradio==4.44.1
streamlit==1.31.0